+++ index.js
@@ -4,11 +4,13 @@ const Dab = require('@rappopo/dab'),
   DabCouch = require('@rappopo/dab-couch'),
   DabEs = require('@rappopo/dab-es')
 
+const delay = (time) => (result) => new Promise(resolve => setTimeout(() => resolve(result), time))
+
 class DabCouchEs extends Dab {
   constructor (options) {
     super(options)
-    this.rekCouch = null
-    this.rekEs = null
+    this.dabCouch = null
+    this.dabEs = null
     this.client = {}
   }
 
@@ -30,57 +32,58 @@ class DabCouchEs extends Dab {
         options: options.options || {
           refresh: 'true'
         }
-      }
+      },
+      delay: options.delay || 1000
     }))
   }
 
   setClient (params) {
     if (!this._.isEmpty(this.client)) return
     const defOptions = this._.omit(this.options, ['couch', 'es'])
-    this.rekCouch = new DabCouch(this._.merge(defOptions, this.options.couch))
-    this.rekEs = new DabEs(this._.merge(defOptions, this.options.es))
-    this.client.rekCouch = this.rekCouch.client
-    this.client.rekEs = this.rekEs.client
+    this.dabCouch = new DabCouch(this._.merge(defOptions, this.options.couch))
+    this.dabEs = new DabEs(this._.merge(defOptions, this.options.es))
+    this.client.dabCouch = this.dabCouch.client
+    this.client.dabEs = this.dabEs.client
   }
 
   find (params) {
     this.setClient(params)
-    return this.rekEs.find(params)
+    return this.dabEs.find(params)
   }
 
   findOne (id, params) {
     this.setClient(params)
-    return this.rekCouch.findOne(id, params)
+    return this.dabCouch.findOne(id, params)
   }
 
   create (body, params) {
     this.setClient(params)
-    return this.rekCouch.create(body, params)
+    return this.dabCouch.create(body, params).then(delay(this.options.delay))
   }
 
   update (id, body, params) {
     this.setClient(params)
-    return this.rekCouch.update(id, body, params)
+    return this.dabCouch.update(id, body, params).then(delay(this.options.delay))
   }
 
   remove (id, params) {
     this.setClient(params)
-    return this.rekCouch.remove(id, params)
+    return this.dabCouch.remove(id, params).then(delay(this.options.delay))
   }
 
   bulkCreate (body, params) {
     this.setClient(params)
-    return this.rekCouch.bulkCreate(body, params)
+    return this.dabCouch.bulkCreate(body, params).then(delay(this.options.delay))
   }
 
   bulkUpdate (body, params) {
     this.setClient(params)
-    return this.rekCouch.bulkUpdate(body, params)
+    return this.dabCouch.bulkUpdate(body, params).then(delay(this.options.delay))
   }
 
   bulkRemove (body, params) {
     this.setClient(params)
-    return this.rekCouch.bulkRemove(body, params)
+    return this.dabCouch.bulkRemove(body, params).then(delay(this.options.delay))
   }
 
 }