+++ test/06-bulkCreate.js
@@ -20,6 +20,8 @@ describe('bulkCreate', function () {
     })
   })
 
+  this.timeout(lib.timeout)
+
   it('should return error if body isn\'t an array', function () {
     const cls = new Cls(lib.options)
     return expect(cls.bulkCreate({ name: 'Rambo' })).to.be.rejectedWith('Require array')
@@ -27,13 +29,13 @@ describe('bulkCreate', function () {
 
   it('should return the correct bulk status', function() {
     const cls = new Cls(lib.options)
-    let p = cls.bulkCreate(lib.bulkDocs)
+    let p = cls.bulkCreate(lib.bulkDocs, { withDetail: true })
     return Promise.all([
       expect(p).to.eventually.have.property('stat').that.have.property('ok').equal(2),
       expect(p).to.eventually.have.property('stat').that.have.property('fail').equal(1),
       expect(p).to.eventually.have.property('stat').that.have.property('total').equal(3),
-      expect(p).to.eventually.have.property('data').that.containSubset([{ _id: 'jack-bauer', message: 'Exists', success: false }]),
-      expect(p).to.eventually.have.property('data').that.containSubset([{ _id: 'johnny-english', success: true }])
+      expect(p).to.eventually.have.property('detail').that.containSubset([{ _id: 'jack-bauer', message: 'Exists', success: false }]),
+      expect(p).to.eventually.have.property('detail').that.containSubset([{ _id: 'johnny-english', success: true }])
     ])
   })