+++ test/08-bulkRemove.js
@@ -11,7 +11,7 @@ chai.use(chaiAsPromised)
 const Cls = require('../index'),
   lib = require('./_lib')
 
-describe('bulkUpdate', function () {
+describe('bulkRemove', function () {
   beforeEach(function (done) {
     this.timeout(lib.timeout)
     lib.resetDb(function (err) {
@@ -20,6 +20,8 @@ describe('bulkUpdate', function () {
     })
   })
 
+  this.timeout(lib.timeout)
+
   it('should return error if body isn\'t an array', function () {
     const cls = new Cls(lib.options)
     return expect(cls.bulkCreate({ name: 'Rambo' })).to.be.rejectedWith('Require array')
@@ -27,15 +29,13 @@ describe('bulkUpdate', function () {
 
   it('should return the correct bulk status', function() {
     const cls = new Cls(lib.options)
-    let docs = lib._.cloneDeep(lib.bulkDocs)
-    docs[0].name = 'Jackie Bauer'
-    let p = cls.bulkUpdate(docs)
+    let p = cls.bulkRemove(lib._.map(lib.bulkDocs, '_id'), { withDetail: true })
     return Promise.all([
       expect(p).to.eventually.have.property('stat').that.have.property('ok').equal(1),
       expect(p).to.eventually.have.property('stat').that.have.property('fail').equal(2),
       expect(p).to.eventually.have.property('stat').that.have.property('total').equal(3),
-      expect(p).to.eventually.have.property('data').that.containSubset([{ _id: 'jack-bauer', success: true }]),
-      expect(p).to.eventually.have.property('data').that.containSubset([{ _id: 'johnny-english', message: 'Not found' }])
+      expect(p).to.eventually.have.property('detail').that.containSubset([{ _id: 'jack-bauer', success: true }]),
+      expect(p).to.eventually.have.property('detail').that.containSubset([{ _id: 'johnny-english', message: 'Not found' }])
     ])
   })